cleanup: refactor to disable metrics
