+++ pkg/apis/config/featureflags_validation.go
@@ -1,3 +1,5 @@
+//go:build !disable_tls
+
 /*
 Copyright 2021 The Tekton Authors
