+++ pkg/substitution/substitution.go
@@ -1,3 +1,5 @@
+//go:build !disable_tls
+
 /*
 Copyright 2019 The Tekton Authors
 
@@ -304,45 +306,6 @@ func matchGroups(matches []string, pattern *regexp.Regexp) map[string]string {
 		groups[name] = matches[i+1]
 	}
 	return groups
-}
-
-// ApplyReplacements returns a string with references to parameters replaced,
-// based on the mapping provided in replacements.
-// For example, if the input string is "foo: $(params.foo)", and replacements maps "params.foo" to "bar",
-// the output would be "foo: bar".
-func ApplyReplacements(in string, replacements map[string]string) string {
-	replacementsList := []string{}
-	for k, v := range replacements {
-		replacementsList = append(replacementsList, fmt.Sprintf("$(%s)", k), v)
-	}
-	// strings.Replacer does all replacements in one pass, preventing multiple replacements
-	// See #2093 for an explanation on why we need to do this.
-	replacer := strings.NewReplacer(replacementsList...)
-	return replacer.Replace(in)
-}
-
-// ApplyArrayReplacements takes an input string, and output an array of strings related to possible arrayReplacements. If there aren't any
-// areas where the input can be split up via arrayReplacements, then just return an array with a single element,
-// which is ApplyReplacements(in, replacements).
-func ApplyArrayReplacements(in string, stringReplacements map[string]string, arrayReplacements map[string][]string) []string {
-	for k, v := range arrayReplacements {
-		stringToReplace := fmt.Sprintf("$(%s)", k)
-
-		// If the input string matches a replacement's key (without padding characters), return the corresponding array.
-		// Note that the webhook should prevent all instances where this could evaluate to false.
-		if (strings.Count(in, stringToReplace) == 1) && len(in) == len(stringToReplace) {
-			return v
-		}
-
-		// same replace logic for star array expressions
-		starStringtoReplace := fmt.Sprintf("$(%s[*])", k)
-		if (strings.Count(in, starStringtoReplace) == 1) && len(in) == len(starStringtoReplace) {
-			return v
-		}
-	}
-
-	// Otherwise return a size-1 array containing the input string with standard stringReplacements applied.
-	return []string{ApplyReplacements(in, stringReplacements)}
 }
 
 // TrimArrayIndex replaces all `[i]` and `[*]` to "".
